+++ domain-order/src/main/java/co/dalicious/domain/order/repository/QOrderCartRepository.java
@@ -16,7 +16,7 @@ public class QOrderCartRepository {
     public final JPAQueryFactory queryFactory;
 
 
-    public Integer getCartId(Integer id) {
+    public BigInteger getCartId(BigInteger id) {
         return queryFactory
                 .select(orderCart.id)
                 .from(orderCart)
